Improved handling of static data
