+++ neural_lam/weather_dataset.py
@@ -1,81 +1,7 @@
-# Standard library
-import os
-
-# Third-party
-import cartopy.crs as ccrs
 import pytorch_lightning as pl
 import torch
-import xarray as xr
-import yaml
 
-
-class ConfigLoader:
-    """
-    Class for loading configuration files.
-
-    This class loads a YAML configuration file and provides a way to access
-    its values as attributes.
-    """
-
-    def __init__(self, config_path, values=None):
-        self.config_path = config_path
-        if values is None:
-            self.values = self.load_config()
-        else:
-            self.values = values
-
-    def load_config(self):
-        """Load configuration file."""
-        with open(self.config_path, encoding="utf-8", mode="r") as file:
-            return yaml.safe_load(file)
-
-    def __getattr__(self, name):
-        keys = name.split(".")
-        value = self.values
-        for key in keys:
-            if key in value:
-                value = value[key]
-            else:
-                return None
-        if isinstance(value, dict):
-            return ConfigLoader(None, values=value)
-        return value
-
-    def __getitem__(self, key):
-        value = self.values[key]
-        if isinstance(value, dict):
-            return ConfigLoader(None, values=value)
-        return value
-
-    def __contains__(self, key):
-        return key in self.values
-
-    def param_names(self):
-        """Return parameter names."""
-        return (
-            self.values["state"]["surface"] + self.values["state"]["atmosphere"]
-        )
-
-    def param_units(self):
-        """Return parameter units."""
-        return (
-            self.values["state"]["surface_units"]
-            + self.values["state"]["atmosphere_units"]
-        )
-
-    def num_data_vars(self, key):
-        """Return the number of data variables for a given key."""
-        surface_vars = len(self.values[key]["surface"])
-        atmosphere_vars = len(self.values[key]["atmosphere"])
-        levels = len(self.values[key]["levels"])
-        return surface_vars + atmosphere_vars * levels
-
-    def projection(self):
-        """Return the projection."""
-        proj_config = self.values["projections"]["class"]
-        proj_class = getattr(ccrs, proj_config["proj_class"])
-        proj_params = proj_config["proj_params"]
-        return proj_class(**proj_params)
+from neural_lam import utils
 
 
 class WeatherDataset(torch.utils.data.Dataset):
@@ -86,65 +12,6 @@ class WeatherDataset(torch.utils.data.Dataset):
     provided configuration. It supports splitting the data into train,
     validation, and test sets.
     """
-
-    def process_dataset(self, dataset_name):
-        """Process a single dataset specified by the dataset name."""
-
-        dataset_path = self.config_loader.zarrs[dataset_name].path
-        if dataset_path is None or not os.path.exists(dataset_path):
-            print(f"Dataset '{dataset_name}' not found at path: {dataset_path}")
-            return None
-        dataset = xr.open_zarr(dataset_path, consolidated=True)
-
-        start, end = (
-            self.config_loader.splits[self.split].start,
-            self.config_loader.splits[self.split].end,
-        )
-        dataset = dataset.sel(time=slice(start, end))
-        dataset = dataset.rename_dims(
-            {
-                v: k
-                for k, v in self.config_loader.zarrs[
-                    dataset_name
-                ].dims.values.items()
-                if k not in dataset.dims
-            }
-        )
-        if "grid" not in dataset.dims:
-            dataset = dataset.stack(grid=("x", "y"))
-
-        vars_surface = []
-        if self.config_loader[dataset_name].surface:
-            vars_surface = dataset[self.config_loader[dataset_name].surface]
-
-        vars_atmosphere = []
-        if self.config_loader[dataset_name].atmosphere:
-            vars_atmosphere = xr.merge(
-                [
-                    dataset[var]
-                    .sel(level=level, drop=True)
-                    .rename(f"{var}_{level}")
-                    for var in self.config_loader[dataset_name].atmosphere
-                    for level in self.config_loader[dataset_name].levels
-                ]
-            )
-
-        if vars_surface and vars_atmosphere:
-            dataset = xr.merge([vars_surface, vars_atmosphere])
-        elif vars_surface:
-            dataset = vars_surface
-        elif vars_atmosphere:
-            dataset = vars_atmosphere
-        else:
-            print("No variables found in dataset {dataset_name}")
-            return None
-
-        dataset = dataset.squeeze(drop=True).to_array()
-        if "time" in dataset.dims:
-            dataset = dataset.transpose("time", "grid", "variable")
-        else:
-            dataset = dataset.transpose("grid", "variable")
-        return dataset
 
     def __init__(
         self,
@@ -166,19 +33,12 @@ class WeatherDataset(torch.utils.data.Dataset):
         self.batch_size = batch_size
         self.ar_steps = ar_steps
         self.control_only = control_only
-        self.config_loader = ConfigLoader(data_config)
+        self.config_loader = utils.ConfigLoader(data_config)
 
-        self.state = self.process_dataset("state")
+        self.state = self.config_loader("state", self.split)
         assert self.state is not None, "State dataset not found"
-        self.static = self.process_dataset("static")
-        self.forcings = self.process_dataset("forcing")
-        self.boundary = self.process_dataset("boundary")
-
-        if self.static is not None:
-            self.static = self.static.expand_dims(
-                {"time": self.state.time}, axis=0
-            )
-            self.state = xr.concat([self.state, self.static], dim="variable")
+        self.forcings = self.config_loader("forcing", self.split)
+        self.boundary = self.config_loader("boundary", self.split)
 
     def __len__(self):
         return len(self.state.time) - self.ar_steps
